requirements.txt
anthropic>=0.52.2
gradio>=5.33
google-genai>=1.0.0
google-generativeai
langchain-google-genai>=0.3.0
langchain>=0.3.0
langchain-anthropic>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.3.0
langchain-tavily>=0.2.0
langchain-together>=0.3.0
langgraph>=0.4.0
requests>=2.32.3
beautifulsoup4>=4.11.0
duckduckgo-search>=3.8.0
wikipedia>=1.4.0
tavily-python>=0.7.0
python-dotenv>=1.0.0
pydantic>=2.0.0
aiohttp>=3.8.0
lxml>=4.9.0
openai>=1.0.0
httpx>=0.28.1,<1.0.0
websockets>=15.0
retry>=0.9.2
